% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialMetrics.R
\name{CHAOS}
\alias{CHAOS}
\title{Calculate CHAOS score}
\usage{
CHAOS(label, location, BNPARAM = NULL)
}
\arguments{
\item{label}{Cluster labels.}

\item{location}{A numeric data matrix containing location information, where
rows are points and columns are location dimensions.}

\item{BNPARAM}{A \code{BiocNeighborParam} object specifying the algorithm to use.
If NULL, automatically deciding from the size of the dataset.}
}
\value{
A numeric value for CHAOS score.
}
\description{
CHAOS score measures the clustering performance by calculating
the mean length of the graph edges in the 1-nearest neighbor (1NN) graph
for each cluster, averaged across clusters.
Lower CHAOS score indicates better spatial domain clustering performance.
}
