% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.check_duplicated_samples}
\alias{.check_duplicated_samples}
\title{Check Duplicated Samples}
\usage{
.check_duplicated_samples(X, threshold = 1e-09)
}
\arguments{
\item{X}{Numeric matrix of samples.}

\item{threshold}{Numeric, the distance threshold to consider samples as duplicates.}
}
\value{
None
}
\description{
Checks for duplicated samples in matrix \code{X}.
}
