% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.compute_cluster_core_distance}
\alias{.compute_cluster_core_distance}
\title{Compute Cluster Core Distance}
\usage{
.compute_cluster_core_distance(dists, d)
}
\arguments{
\item{dists}{Numeric matrix of distances.}

\item{d}{Integer, the dimensionality.}
}
\value{
Numeric vector of core distances for each point.
}
\description{
Computes the core distance for each point in a cluster.
}
