% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.compute_cross_dists}
\alias{.compute_cross_dists}
\title{Compute Cross Distances}
\usage{
.compute_cross_dists(X, inds_a, inds_b, distance)
}
\arguments{
\item{X}{Numeric matrix of samples.}

\item{inds_a}{Integer vector of indices for the first set.}

\item{inds_b}{Integer vector of indices for the second set.}

\item{distance}{String specifying the distance metric.}
}
\value{
Numeric matrix of cross distances.
}
\description{
Computes the cross distances between two sets of indices in matrix \code{X}.
}
\keyword{internal}
