% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.compute_mutual_reach_dists}
\alias{.compute_mutual_reach_dists}
\title{Compute Mutual Reachability Distances}
\usage{
.compute_mutual_reach_dists(dists, d)
}
\arguments{
\item{dists}{Numeric matrix of distances.}

\item{d}{Float, the dimensionality.}
}
\value{
A list containing core distances and mutual reachability distances.
}
\description{
Computes the mutual reachability distances between points.
}
