% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.compute_pair_to_pair_dists}
\alias{.compute_pair_to_pair_dists}
\title{Compute Pair to Pair Distances}
\usage{
.compute_pair_to_pair_dists(X, metric = "euclidean")
}
\arguments{
\item{X}{Numeric matrix.}

\item{metric}{String specifying the metric to compute the distances.}
}
\value{
Numeric matrix of pairwise distances with self-distances set to \code{Inf}.
}
\description{
Compute the pairwise distances between points in matrix \code{X}.
}
