% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.convert_singleton_clusters_to_noise}
\alias{.convert_singleton_clusters_to_noise}
\title{Convert Singleton Clusters to Noise}
\usage{
.convert_singleton_clusters_to_noise(labels, noise_id)
}
\arguments{
\item{labels}{Integer vector of cluster IDs.}

\item{noise_id}{Integer, the ID for noise.}
}
\value{
Integer vector with singleton clusters converted to noise.
}
\description{
Converts clusters containing a single instance to noise.
}
