% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{.FMeasure}
\alias{.FMeasure}
\title{F measure}
\usage{
.FMeasure(true, pred, silent = TRUE)
}
\arguments{
\item{true}{Array containing real cluster labels for each sample}

\item{pred}{Array containing predicted cluster labels for each
sample}

\item{silent}{Logical, if FALSE, print some information about
precision and recall}
}
\value{
F measure score
}
\description{
Compute the F measure between two clustering results. This is directly copied
from the package \code{FlowSOM}.
}
\examples{
# Generate some random data as an example
true <- sample(1:5,100,replace = TRUE)
pred <- sample(1:6, 100, replace = TRUE)

# Calculate the FMeasure
FMeasure(true,pred)
}
