% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.fn_density_separation}
\alias{.fn_density_separation}
\title{Density Separation of a Pair of Clusters}
\usage{
.fn_density_separation(
  cls_i,
  cls_j,
  dists,
  internal_core_dists_i,
  internal_core_dists_j
)
}
\arguments{
\item{cls_i}{Integer, first cluster index.}

\item{cls_j}{Integer, second cluster index.}

\item{dists}{Numeric matrix of distances.}

\item{internal_core_dists_i}{Numeric vector of core distances for cluster i.}

\item{internal_core_dists_j}{Numeric vector of core distances for cluster j.}
}
\value{
A list containing the cluster indices and their density separation.
}
\description{
Computes the density separation between two clusters.
}
\keyword{internal}
