% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.get_submatrix}
\alias{.get_submatrix}
\title{Get Sub matrix}
\usage{
.get_submatrix(arr, inds_a = NULL, inds_b = NULL)
}
\arguments{
\item{arr}{Numeric matrix.}

\item{inds_a}{Optional integer vector for row indices.}

\item{inds_b}{Optional integer vector for column indices.}
}
\value{
Numeric matrix representing the sub matrix.
}
\description{
Extract a sub matrix from a matrix based on optional row and column indices.
}
\keyword{internal}
