% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{findSpatialKNN}
\alias{findSpatialKNN}
\title{Find the k nearest spatial neighbors}
\usage{
findSpatialKNN(location, k, keep_ties = TRUE, n = 5, BNPARAM = NULL, ...)
}
\arguments{
\item{location}{A numeric data matrix containing location information, where
rows are points and columns are location dimensions.}

\item{k}{The number of nearest neighbors to look at.}

\item{keep_ties}{A Boolean indicating if ties of neighbors are counted once
or not. If TRUE, neighbors of the same distances will be counted only once,
and the resulting KNN will be more than k if there are any ties exist.}

\item{n}{When ties are kept, maximally \code{k*n} neighbors will be looked at.}

\item{BNPARAM}{A \code{BiocNeighborParam} object specifying the algorithm to use.
If NULL, automatically deciding from the size of the dataset.}
}
\value{
return a list containing the indices of knn for each object.
}
\description{
For a given dataset, find the \code{k} nearest neighbors for each
object based on their spatial locations, with the option of handling ties.
}
