% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyPartitionMetrics.R
\name{fuzzyHardMetrics}
\alias{fuzzyHardMetrics}
\title{fuzzyHardMetrics}
\usage{
fuzzyHardMetrics(
  hardPred,
  hardTruth,
  fuzzyTruth,
  nperms = NULL,
  verbose = TRUE,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{hardPred}{An atomic vector coercible to a factor or integer vector
containing the predicted hard labels.}

\item{hardTruth}{An atomic vector coercible to a factor or integer vector
containing the true hard labels. Must have the same length as \code{hardPred}.}

\item{fuzzyTruth}{A object coercible to a numeric matrix with membership
probability of elements (rows) in clusters (columns). Must have the same
number of rows as the length of \code{hardTruth}.}

\item{nperms}{The number of permutations (for correction for chance). If
NULL (default), a first set of 10 permutations will be run to estimate
whether the variation across permutations is above 0.0025, in which case
more (max 1000) permutations will be run.}

\item{verbose}{Logical; whether to print info and warnings, including the
standard error of the mean across permutations (giving an idea of the
precision of the adjusted metrics).}

\item{BPPARAM}{BiocParallel params for multithreading (default none)}
}
\value{
A list of metrics:
\item{NDC}{Hullermeier's NDC (fuzzy rand index)}
\item{ACI}{Ambrosio's Adjusted Concordance Index (ACI), i.e. a
permutation-based fuzzy version of the adjusted Rand index.}
\item{fuzzyWH}{Fuzzy Wallace Homogeneity index}
\item{fuzzyWC}{Fuzzy Wallace Completeness index}
\item{fuzzyAWH}{Adjusted fuzzy Wallace Homogeneity index}
\item{fuzzyAWC}{Adjusted fuzzy Wallace Completeness index}
}
\description{
Computes fuzzy-hard versions of pair-sorting partition metrics to compare a
hard clustering with both a fuzzy and hard truth. This was especially
designed for cases where the fuzzy truth represents an uncertainty of a hard
truth. Briefly put, the maximum of the pair concordance between the
clustering and either the hard or the fuzzy truth is used, and the hard truth
is used to compute completeness. See \code{\link{fuzzyPartitionMetrics}} for
the more standard implementation of the metrics.
}
\examples{
# generate a fuzzy truth:
fuzzyTruth <- matrix(c(
  0.95, 0.025, 0.025, 
  0.98, 0.01, 0.01, 
  0.96, 0.02, 0.02, 
  0.95, 0.04, 0.01, 
  0.95, 0.01, 0.04, 
  0.99, 0.005, 0.005, 
  0.025, 0.95, 0.025, 
  0.97, 0.02, 0.01, 
  0.025, 0.025, 0.95), 
  ncol = 3, byrow=TRUE)
# a hard truth:
hardTruth <- apply(fuzzyTruth,1,FUN=which.max)
# some predicted labels:
hardPred <- c(1,1,1,1,1,1,2,2,2)
fuzzyHardMetrics(hardPred, hardTruth, fuzzyTruth, nperms=3)
}
\references{
Hullermeier et al. 2012; 10.1109/TFUZZ.2011.2179303;

D'Ambrosio et al. 2021; 10.1007/s00357-020-09367-0
}
\seealso{
\code{\link[=fuzzyPartitionMetrics]{fuzzyPartitionMetrics()}}
}
\author{
Pierre-Luc Germain
}
