% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyPartitionMetrics.R
\name{fuzzySpotAgreement}
\alias{fuzzySpotAgreement}
\title{fuzzySpotAgreement}
\usage{
fuzzySpotAgreement(P, Q)
}
\arguments{
\item{P}{A object coercible to a numeric matrix with membership probability
of elements (rows) in clusters (columns)}

\item{Q}{A object coercible to a numeric matrix with membership probability
of elements (rows) in clusters (columns). Must have the same number of rows
as \code{P}}
}
\value{
A numeric vector of agreement scores for each row of \code{P}.
}
\description{
Per-spot agreement between two fuzzy partitionings
}
\examples{
# generate fuzzy partitions:
m1 <- matrix(c(0.95, 0.025, 0.025, 
               0.98, 0.01, 0.01, 
               0.96, 0.02, 0.02, 
               0.95, 0.04, 0.01, 
               0.95, 0.01, 0.04, 
               0.99, 0.005, 0.005, 
               0.025, 0.95, 0.025, 
               0.97, 0.02, 0.01, 
               0.025, 0.025, 0.95), 
               ncol = 3, byrow=TRUE)
m2 <- matrix(c(0.95, 0.025, 0.025,  
               0.98, 0.01, 0.01, 
               0.96, 0.02, 0.02, 
               0.025, 0.95, 0.025, 
               0.02, 0.96, 0.02, 
               0.01, 0.98, 0.01, 
               0.05, 0.05, 0.95, 
               0.02, 0.02, 0.96, 
               0.01, 0.01, 0.98), 
               ncol = 3, byrow=TRUE)
colnames(m1) <- colnames(m2) <- LETTERS[1:3]
fuzzySpotAgreement(m1,m2)
}
