% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialMetrics.R
\name{getAgreement}
\alias{getAgreement}
\title{getAgreement}
\usage{
getAgreement(pred, true, usePairs = TRUE)
}
\arguments{
\item{pred}{A vector of predicted clusters}

\item{true}{A vector of true class labels}

\item{usePairs}{Logical; whether to compute over pairs instead of elements}
}
\value{
A vector of agreement scores
}
\description{
Per-spot agreement between a clustering and a ground truth
}
\examples{
# Give the SPE:
spe$agreement <- getAgreement(spe$BayesSpace_default, true=spe$ground_truth)
}
