% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmbeddingMetrics.R
\name{getEmbeddingGlobalMetrics}
\alias{getEmbeddingGlobalMetrics}
\title{getEmbeddingGlobalMetrics}
\usage{
getEmbeddingGlobalMetrics(
  x,
  labels,
  metrics = c("meanSW", "meanClassSW", "pnSW", "minClassSW", "cdbw", "cohesion",
    "compactness", "sep", "dbcv"),
  distance = "euclidean",
  ...
)
}
\arguments{
\item{x}{A data.frame or matrix (with features as columns and items as rows)
from which the metrics will be computed.}

\item{labels}{A vector containing the labels of the predicted clusters. Must
be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{metrics}{The metrics to compute.}

\item{distance}{The distance metric to use (default euclidean).}

\item{...}{Optional arguments. See details.}
}
\value{
A data.frame (with 1 row) of metrics.
}
\description{
Computes dataset-level, embedding-based metrics.
}
