% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFuzzyPartitionMetrics.R
\name{getFuzzyPartitionElementMetrics}
\alias{getFuzzyPartitionElementMetrics}
\title{getFuzzyPartitionElementMetrics}
\usage{
getFuzzyPartitionElementMetrics(
  hardTrue = NULL,
  fuzzyTrue = NULL,
  hardPred = NULL,
  fuzzyPred = NULL,
  fuzzy_true = TRUE,
  fuzzy_pred = FALSE,
  metrics = c("spotAgreement"),
  useNegatives = TRUE,
  verbose = TRUE,
  usePairs = TRUE
)
}
\arguments{
\item{hardTrue}{A vector of true cluster labels}

\item{fuzzyTrue}{A object coercible to a numeric matrix with membership
probability of elements (rows) in clusters (columns). Must have the same
number of rows as the length of \code{hardTrue}.}

\item{hardPred}{A vector of predicted cluster labels}

\item{fuzzyPred}{A object coercible to a numeric matrix with membership
probability of elements (rows) in clusters (columns).}

\item{fuzzy_true}{Logical; whether the truth is fuzzy.}

\item{fuzzy_pred}{Logical; whether the prediction is fuzzy.}

\item{metrics}{The metrics to compute. Currently only \code{"spotAgreement"} is included at the element level.}

\item{useNegatives}{Logical; whether to include negative pairs in the
agreement score (tends to result in a larger overall agreement and lower
dynamic range of the score). Default TRUE.}

\item{verbose}{Logical; whether to print expected memory usage for large
datasets.}

\item{usePairs}{Logical; whether to compute over pairs instead of elements.
Only useful when \code{fuzzy_true=TRUE} and \code{fuzzy_pred=FALSE}.}
}
\value{
A dataframe of metric values.
}
\description{
Computes a selection of external fuzzy clustering evaluation metrics at the element level.
}
\examples{
# generate fuzzy partitions:
m1 <- matrix(c(0.95, 0.025, 0.025, 
               0.98, 0.01, 0.01, 
               0.96, 0.02, 0.02, 
               0.95, 0.04, 0.01, 
               0.95, 0.01, 0.04, 
               0.99, 0.005, 0.005, 
               0.025, 0.95, 0.025, 
               0.97, 0.02, 0.01, 
               0.025, 0.025, 0.95), 
               ncol = 3, byrow=TRUE)
m2 <- matrix(c(0.95, 0.025, 0.025,  
               0.98, 0.01, 0.01, 
               0.96, 0.02, 0.02, 
               0.025, 0.95, 0.025, 
               0.02, 0.96, 0.02, 
               0.01, 0.98, 0.01, 
               0.05, 0.05, 0.95, 
               0.02, 0.02, 0.96, 
               0.01, 0.01, 0.98), 
               ncol = 3, byrow=TRUE)
colnames(m1) <- colnames(m2) <- LETTERS[1:3]
getFuzzyPartitionElementMetrics(fuzzyTrue=m1,fuzzyPred=m2, fuzzy_true=TRUE, fuzzy_pred=TRUE)
}
