% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphClassMetrics.R
\name{getGraphClassMetrics}
\alias{getGraphClassMetrics}
\title{getGraphClassMetrics}
\usage{
getGraphClassMetrics(
  x,
  labels,
  metrics = c("SI", "NP", "AMSP", "PWC", "NCE"),
  ...
)
}
\arguments{
\item{x}{Either an igraph object, or a list of nearest neighbors (see details
below), or a data.frame or matrix (with features as columns and items as
rows) from which nearest neighbors will be computed.}

\item{labels}{Either a factor or a character vector indicating the true class
label of each element (i.e. row or vertex) of \code{x}.}

\item{metrics}{The metrics to compute. If omitted, main metrics will be
computed.}

\item{...}{Passed to other methods}

\item{k}{The number of nearest neighbors to compute and/or use. Can be
omitted if \code{x} is a graph or list of nearest neighbors.}

\item{BNPARAM}{A BiocNeighbors parameter object to compute kNNs. Ignored
unless the input is a matrix or data.frame. If omitted, the Annoy
approximation will be used if there are more than 500 elements.}
}
\value{
A data.frame of metrics for each class.
}
\description{
Short description...
}
\details{
Additional details...
}
\examples{
# generate random data and labels:
dat <- matrix(rnorm(300), nrow=30)
dat[1:10,1:3] <- dat[1:10,1:3]+2
labels <- rep(LETTERS[1:3], each=10)
# compute graph metrics:
getGraphClassMetrics(dat, labels, k=4)
}
