% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialMetrics.R
\name{getNeihboringPairAgreement}
\alias{getNeihboringPairAgreement}
\title{getNeihboringPairAgreement}
\usage{
getNeihboringPairAgreement(
  true,
  pred,
  location,
  k = 20L,
  useNegatives = FALSE,
  distWeights = TRUE,
  BNPARAM = NULL
)
}
\arguments{
\item{true}{A vector of true class labels}

\item{pred}{A vector of predicted clusters}

\item{location}{A matrix or data.frame with spatial dimensions as columns.
Alternatively, a nearest neighbor object as produced by
\code{\link[BiocNeighbors]{findKNN}}.}

\item{k}{Approximate number of nearest neighbors to consider}

\item{useNegatives}{Logical; whether to include the consistency of negative
pairs in the score (default FALSE).}

\item{distWeights}{Logical; whether to weight agreement by distance (default
TRUE).}

\item{BNPARAM}{A BiocNeighbors parameter object to compute kNNs. Ignored
unless the input is a matrix or data.frame. If omitted, the Annoy
approximation will be used if there are more than 500 elements.}
}
\value{
A vector of agreement scores
}
\description{
Per-spot local agreement between a clustering and a ground truth
}
