% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartitionMetrics.R
\name{getPartitionElementMetrics}
\alias{getPartitionElementMetrics}
\title{getPartitionElementMetrics}
\usage{
getPartitionElementMetrics(
  true,
  pred,
  metrics = c("ASPC"),
  usePairs = TRUE,
  useNegatives = TRUE
)
}
\arguments{
\item{true}{A vector of true class labels}

\item{pred}{A vector of predicted clusters}

\item{metrics}{The metrics to compute.}

\item{usePairs}{Logical; whether to compute over pairs instead of elements
Recommended and TRUE by default.}

\item{useNegatives}{Logical; whether to include the consistency of negative
pairs in the score (default FALSE).}
}
\value{
A dataframe of metrics.
}
\description{
Computes a selection of external evaluation metrics for partition. The
metrics are reported per element.
}
\keyword{internal}
