% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartitionMetrics.R
\name{getPartitionMetrics}
\alias{getPartitionMetrics}
\title{getPartitionMetrics}
\usage{
getPartitionMetrics(
  true,
  pred,
  metrics = c("RI", "WC", "WH", "ARI", "NCR", "AWC", "AWH", "MI", "AMI", "VI", "EH",
    "EC", "VM", "FM", "VDM", "Mirkin", "MHM", "MMM"),
  ...
)
}
\arguments{
\item{true}{A vector containing the labels of the true classes. Must be a
vector of characters, integers, numerics, or a factor, but not a list.}

\item{pred}{A vector containing the labels of the predicted clusters. Must
be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{metrics}{The metrics to compute. If omitted, main metrics will be
computed. RI: Rand Index; WC: Wallace Completeness; WH: Wallace
Homogeneity; ARI:Adjusted Rand Index; AWC: Adjusted Wallace Completeness;
AWH: Adjusted Wallace Homogeneity; NCR: Normalized class size Rand index;
MI: Mutual Information; AMI: Adjusted Mutual Information;
VI: Variation of Information; EH: (Entropy-based) Homogeneity;
EC: (Entropy-based) Completeness; VM: V-measure;
FM: F-measure/weighted average F1 score; VDM: Van Dongen Measure;
MHM: Meila-Heckerman Measure; MMM: Maximum-Match Measure;
Mirkin: Mirkin Metric.}
}
\value{
A list of metric results.
}
\description{
Computes a selection of external evaluation metrics for partition. The
metrics are reported per dataset.
}
\details{
Additional details...
}
\examples{
true <- rep(LETTERS[1:3], each=10)
pred <- c(rep("A", 8), rep("B", 9), rep("C", 3), rep("D", 10))
getPartitionMetrics(true, pred)
}
