% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialExternalMetrics.R
\name{getSpatialExternalMetrics}
\alias{getSpatialExternalMetrics}
\title{getSpatialExternalMetrics}
\usage{
getSpatialExternalMetrics(
  true,
  pred,
  location,
  k = 6,
  alpha = 0.5,
  level = "class",
  metrics = c("SpatialWH", "SpatialAWH", "SpatialWC", "SpatialAWC"),
  fuzzy_true = TRUE,
  fuzzy_pred = FALSE,
  ...
)
}
\arguments{
\item{true}{A vector containing the labels of the true classes. Must be a
vector of characters, integers, numerics, or a factor, but not a list.}

\item{pred}{A vector containing the labels of the predicted clusters. Must
be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{location}{A matrix or data.frame of coordinates}

\item{k}{The number of neighbors used when calculating the fuzzy
class memberships for fuzzy metrics, or when calculating the weighted
accuracy.}

\item{alpha}{The parameter to control to what extend the spot itself contribute
to the class composition calculation. \code{"equal"} means it is weighted the
same as other neighbors. A numeric value between \code{0} and \code{1} means the weight of the
frequency contribution for the spot itself, and the frequency contribution
for its knn is then \code{1-alpha}. By default \code{0.5}.}

\item{level}{The level to calculate the metrics. Options include \code{"element"},
\code{"class"} and \code{"global"}.}

\item{metrics}{The metrics to compute. See details.}

\item{fuzzy_true}{Logical. Indicating if the \code{true} label will be transformed
into fuzzy cluster membership representation.}

\item{fuzzy_pred}{Logical. Indicating if the \code{pred} label will be transformed
into fuzzy cluster membership representation.}

\item{...}{Optional params for \code{\link[=getFuzzyPartitionElementMetrics]{getFuzzyPartitionElementMetrics()}} or \code{\link[=findSpatialKNN]{findSpatialKNN()}}.}
}
\value{
A data.frame of metrics.
}
\description{
Computes a selection of external clustering evaluation metrics for spatial
data.
}
\details{
The allowed values for \code{metrics} depend on the value of \code{level}:
\itemize{
\item If \code{level = "element"}, the allowed \code{metrics} are: \code{"spotAgreement"}.
\item If \code{level = "class"}, the allowed \code{metrics} are: \code{"SpatialWH"},\code{"SpatialAWH"}, \code{"SpatialWC"},\code{"SpatialAWC"}.
\item If \code{level = "global"}, the allowed \code{metrics} are: \code{"SpatialRI"},\code{"SpatialARI"},\code{"SpatialWH"},\code{"SpatialAWH"}, \code{"SpatialWC"},\code{"SpatialAWC"},\code{"SpatialAccuracy"}.
}
}
\examples{
data <- sp_toys
getSpatialExternalMetrics(data$label, data$p1, data[,c("x", "y")], k=6, level="class")
}
