% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialMetrics.R
\name{getSpatialGlobalExternalMetrics}
\alias{getSpatialGlobalExternalMetrics}
\title{getSpatiaGloballExternalMetrics}
\usage{
getSpatialGlobalExternalMetrics(
  true,
  pred,
  location,
  k = 6,
  metrics = c("SpatialRI", "SpatialARI", "SpatialWH", "SpatialAWH", "SpatialWC",
    "SpatialAWC", "SpatialAccuracy"),
  ...
)
}
\description{
Computes a selection of external clustering evaluation metrics for spatial
data at the global level.
}
