% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{matchSets}
\alias{matchSets}
\title{Match two partitions using Hungarian algorithm}
\usage{
matchSets(pred, true, forceMatch = TRUE, returnIndices = is.integer(true))
}
\arguments{
\item{pred}{An integer or factor of cluster labels}

\item{true}{An integer or factor of reference labels}

\item{forceMatch}{Logical; whether to enforce a match for every set of \code{pred}}

\item{returnIndices}{Logical; whether to return indices rather than levels}
}
\value{
A vector of matching sets (i.e. level) from \code{true} for every set
(i.e. level) of \code{pred}.
}
\description{
Match sets from a partitions to a reference partition using the Hungarian
algorithm to optimize F1 scores.
}
