% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metric_info}
\alias{metric_info}
\title{Metrics Information}
\format{
\subsection{\code{metric_info}}{

A data frame.
}
}
\usage{
metric_info
}
\description{
A dataframe storing the information of all metrics. The code to
generate the dataset is at
system.file('inst/scripts/', 'metric_info.R', package='poem').
}
\keyword{datasets}
