% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mockData}
\alias{mockData}
\title{mockData}
\usage{
mockData(Ns = c(25, 15), classDiff = 2, nSpread = 2)
}
\arguments{
\item{Ns}{A vector of 2 positive integers specifying the number of elements
of each class.}

\item{classDiff}{The magnitude of the difference between the two classes,
as a proportion of the standard deviation.}

\item{nSpread}{The extent of spread in the first class.}
}
\value{
A data.frame with \code{x} and \code{y} coordinates and a \code{class} column.
}
\description{
Generates mock 2-dimensional data of two classes of points, for testing.
}
\examples{
d <- mockData()
}
