% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialMetrics.R
\name{nnWeightedAccuracy}
\alias{nnWeightedAccuracy}
\title{nnWeightedAccuracy}
\usage{
nnWeightedAccuracy(true, pred, location, k = 5, ...)
}
\arguments{
\item{true}{True class labels (vector coercible to factor)}

\item{pred}{Predicted labels (vector coercible to factor)}

\item{location}{The spatial coordinates to compute the nearest neighbors.}

\item{k}{Number of nearest neighbors}

\item{...}{Optional params passed to \code{\link[=findSpatialKNN]{findSpatialKNN()}}.}
}
\value{
A scalar representing the weighted accuracy.
}
\description{
Computes an accuracy score which weighs elements/spots that are misclassified
by the proportion of their (spatial) neighborhood that is not of the
element/spot's predicted class. This reduces the weight of misclassifications
happening at the boundary between domains.
}
\keyword{internal}
