% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{setMatchingAccuracy}
\alias{setMatchingAccuracy}
\title{The non-spatially-weighted counterpart of nnWeightedAccuracy}
\usage{
setMatchingAccuracy(true, pred)
}
\arguments{
\item{true}{True class labels (vector coercible to factor)}

\item{pred}{Predicted labels (vector coercible to factor)}
}
\value{
A scalar representing the weighted accuracy.
}
\description{
The non-spatially-weighted counterpart of nnWeightedAccuracy
}
\keyword{internal}
