% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_toys}
\alias{sp_toys}
\title{Toy examples of spatial data}
\format{
\subsection{\code{sp_toys}}{

A data frame with 240 rows and 11 columns, representing a 16 x 15 array of spots:
\describe{
\item{x, y}{Coordinates of each spots.}
\item{row, col}{The row and column index of each spots.}
\item{label}{Ground truth labels. Either 1 or 2.}
\item{p1-p6}{Hypothetical predicted spatial clustering labels.}
}
}
}
\usage{
sp_toys
}
\description{
Toy examples of spatial data.
}
\keyword{datasets}
