% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{.fn_density_sparseness}
\alias{.fn_density_sparseness}
\title{Density Sparseness of a Cluster}
\usage{
.fn_density_sparseness(cls_inds, dists, d, use_igraph_mst)
}
\arguments{
\item{cls_inds}{Integer vector of cluster indices.}

\item{dists}{Numeric matrix of distances.}

\item{d}{Integer, the dimensionality.}

\item{use_igraph_mst}{Logical flag to use MST implementation
in igraph. Currently only mst from igraph is implemented.}
}
\value{
A list containing the density sparseness, internal core distances, and internal node indices.
}
\description{
Computes the density sparseness for a given cluster.
}
