% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyPartitionMetrics.R
\name{fuzzyPartitionMetrics}
\alias{fuzzyPartitionMetrics}
\title{fuzzyPartitionMetrics}
\usage{
fuzzyPartitionMetrics(
  P,
  Q,
  computeWallace = TRUE,
  nperms = 1000,
  verbose = TRUE,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{P}{A object coercible to a numeric matrix with membership probability
of elements (rows) in clusters (columns)}

\item{Q}{A object coercible to a numeric matrix with membership probability
of elements (rows) in clusters (columns). Must have the same number of rows
as \code{P}}

\item{computeWallace}{Logical; whether to compute the individual fuzzy
versions of the Wallace indices (increases running time).}

\item{nperms}{The number of permutations (for correction for chance)}

\item{verbose}{Logical; whether to print info and warnings, including the
standard error of the mean across permutations (giving an idea of the
precision of the adjusted metrics).}

\item{BPPARAM}{BiocParallel params for multithreading (default none)}
}
\value{
A list of metrics:
\item NDC : Hullermeier's NDC (fuzzy rand index)
\item ACI : Ambrosio's Adjusted Concordance Index (ACI), i.e. a
permutation-based fuzzy version of the adjusted Rand index.
\item fuzzyWallace1 Fuzzy Wallace index for each partition of \code{Q}
\item fuzzyWallace2 Fuzzy Wallace index for each partition of \code{P}
\item fuzzyAdjW1 Adjusted fuzzy Wallace index for each partition of \code{Q}
\item fuzzyAdjW2 Adjusted fuzzy Wallace index for each partition of \code{P}
}
\description{
Computes fuzzy versions of pair-sorting partition metrics. This is largely
based on the permutation-based implementation by Antonio D'Ambrosio from the
ConsRankClass package, modified to also compute the fuzzy versions of the
adjusted Wallace indices and implement multithreading.
}
\examples{
# generate fuzzy partitions:
m1 <- matrix(c(0.95, 0.025, 0.025, 
               0.98, 0.01, 0.01, 
               0.96, 0.02, 0.02, 
               0.95, 0.04, 0.01, 
               0.95, 0.01, 0.04, 
               0.99, 0.005, 0.005, 
               0.025, 0.95, 0.025, 
               0.97, 0.02, 0.01, 
               0.025, 0.025, 0.95), 
               ncol = 3, byrow=TRUE)
m2 <- matrix(c(0.95, 0.025, 0.025,  
               0.98, 0.01, 0.01, 
               0.96, 0.02, 0.02, 
               0.025, 0.95, 0.025, 
               0.02, 0.96, 0.02, 
               0.01, 0.98, 0.01, 
               0.05, 0.05, 0.95, 
               0.02, 0.02, 0.96, 
               0.01, 0.01, 0.98), 
               ncol = 3, byrow=TRUE)
colnames(m1) <- colnames(m2) <- LETTERS[1:3]
fuzzyPartitionMetrics(m1,m2)
}
\references{
Hullermeier et al. 2012; 10.1109/TFUZZ.2011.2179303;

D’Ambrosio et al. 2021; 10.1007/s00357-020-09367-0
}
\author{
Pierre-Luc Germain
}
