% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyPartitionMetrics.R
\name{fuzzySpotAgreement}
\alias{fuzzySpotAgreement}
\title{fuzzySpotAgreement}
\usage{
fuzzySpotAgreement(P, Q)
}
\arguments{
\item{P}{A object coercible to a numeric matrix with membership probability
of elements (rows) in clusters (columns)}

\item{Q}{A object coercible to a numeric matrix with membership probability
of elements (rows) in clusters (columns). Must have the same number of rows
as \code{P}}
}
\value{
A numeric vector of agreement scores for each row of \code{P}.
}
\description{
Per-spot agreement between two fuzzy partitionings
}
