% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialMetrics.R
\name{getAgreement}
\alias{getAgreement}
\title{getAgreement}
\usage{
getAgreement(true, pred, usePairs = TRUE, useNegatives = FALSE)
}
\arguments{
\item{true}{A vector of true class labels}

\item{pred}{A vector of predicted clusters}

\item{usePairs}{Logical; whether to compute over pairs instead of elements}

\item{useNegatives}{Logical; whether to include the consistency of negative
pairs in the score (default FALSE).}
}
\value{
A vector of agreement scores
}
\description{
Per-spot agreement between a clustering and a ground truth
}
\examples{
true <- rep(1:3,each=3)
pred <- rep(1:2, c(4,5))
getAgreement(true, pred)
}
