% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFuzzyPartitionMetrics.R
\name{getFuzzyPartitionMetrics}
\alias{getFuzzyPartitionMetrics}
\title{getFuzzyPartitionMetrics}
\usage{
getFuzzyPartitionMetrics(true, pred, location, k = 6, alpha = "equal", ...)
}
\arguments{
\item{true}{A vector containing the labels of the true classes. Must be a
vector of characters, integers, numerics, or a factor, but not a list.}

\item{pred}{A vector containing the labels of the predicted clusters. Must
be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{k}{the number of neighbors to consider when transforming the location
information into fuzzy class memberships.}

\item{alpha}{The parameter to control to what extend the spot itself
contribute to the class composition calculation. "equal" means it is
weighted the same as other NNs. A numeric value between 0 and 1 means the
weight of the frequency contribution for the spot itself, and the frequency
contribution for its knn is then 1-alpha.}
}
\value{
A list of metric results.
}
\description{
Computes a selection of external fuzzy clustering evaluation metrics for spatial data.
}
\details{
Additional details...
}
