% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartitionClassMetrics.R
\name{getPartitionClassMetrics}
\alias{getPartitionClassMetrics}
\title{getPartitionClassMetrics}
\usage{
getPartitionClassMetrics(
  true,
  pred,
  metrics = c("WC", "WH", "AWC", "AWH", "FM")
)
}
\arguments{
\item{true}{A vector containing the labels of the true classes. Must be a
vector of characters, integers, numerics, or a factor, but not a list.}

\item{pred}{A vector containing the labels of the predicted clusters. Must
be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{metrics}{The metrics to compute. If omitted, main metrics will be
computed.}
}
\value{
A dataframe of metrics.
}
\description{
Computes a selection of external evaluation metrics for partition. The
metrics are reported per class.
}
