% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartitionMetrics.R
\name{getPartitionMetrics}
\alias{getPartitionMetrics}
\title{getPartitionMetrics}
\usage{
getPartitionMetrics(
  true,
  pred,
  metrics = c("RI", "WC", "WH", "ARI", "AWC", "AWH"),
  ...
)
}
\arguments{
\item{true}{A vector containing the labels of the true classes. Must be a
vector of characters, integers, numerics, or a factor, but not a list.}

\item{pred}{A vector containing the labels of the predicted clusters. Must
be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{metrics}{The metrics to compute. If omitted, main metrics will be
computed.}
}
\value{
A list of metrics.
}
\description{
Short description...
}
\details{
Additional details...
}
\examples{
true <- rep(LETTERS[1:3], each=10)
pred <- c(rep("A", 8), rep("B", 9), rep("C", 3), rep("D", 10))
getPartitionMetrics(true, pred)
}
