% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartitionMetrics.R
\name{getPartitionMetrics}
\alias{getPartitionMetrics}
\title{getPartitionMetrics}
\usage{
getPartitionMetrics(
  true,
  pred,
  metrics = c("RI", "WC", "WH", "ARI", "NCR", "AWC", "AWH", "MI", "AMI", "VI", "EH",
    "EC", "VM", "FM", "VDM", "Mirkin", "MHM", "MMM"),
  ...
)
}
\arguments{
\item{true}{A vector containing the labels of the true classes. Must be a
vector of characters, integers, numerics, or a factor, but not a list.}

\item{pred}{A vector containing the labels of the predicted clusters. Must
be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{metrics}{The metrics to compute. If omitted, main metrics will be
computed. See below for more details.}
}
\value{
A list of metric results. Possible metrics are:
\item{RI}{Rand Index}
\item{WC}{Wallace Completeness}
\item{WH}{Wallace Homogeneity}
\item{ARI}{Adjusted Rand Index}
\item{AWC}{Adjusted Wallace Completeness}
\item{AWH}{Adjusted Wallace Homogeneity}
\item{NCR}{Normalized class size Rand index}
\item{MI}{Mutual Information}
\item{AMI}{Adjusted Mutual Information}
\item{VI}{Variation of Information}
\item{EH}{(Entropy-based) Homogeneity}
\item{EC}{(Entropy-based) Completeness}
\item{VM}{V-measure}
\item{FM}{F-measure/weighted average F1 score}
\item{VDM}{Van Dongen Measure}
\item{MHM}{Meila-Heckerman Measure}
\item{MMM}{Maximum-Match Measure}
\item{Mirkin}{Mirkin Metric}
}
\description{
Computes a selection of external evaluation metrics for partition. The
metrics are reported per dataset.
}
\details{
Additional details...
}
\examples{
true <- rep(LETTERS[1:3], each=10)
pred <- c(rep("A", 8), rep("B", 9), rep("C", 3), rep("D", 10))
getPartitionMetrics(true, pred)
}
