% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialMetrics.R
\name{getSpatialClassExternalMetrics}
\alias{getSpatialClassExternalMetrics}
\title{getSpatialClassExternalMetrics}
\usage{
getSpatialClassExternalMetrics(
  true,
  pred,
  location,
  k = 6,
  metrics = c("SpatialWH", "SpatialAWH", "SpatialWC", "SpatialAWC"),
  ...
)
}
\description{
Computes a selection of external clustering evaluation metrics for spatial
data at the class/cluster level.
}
