% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialMetrics.R
\name{getSpatialClassExternalMetrics}
\alias{getSpatialClassExternalMetrics}
\title{Compute class-level external evaluation metrics for spatially-resolved data}
\usage{
getSpatialClassExternalMetrics(
  true,
  pred,
  location,
  k = 6,
  metrics = c("SpatialWH", "SpatialAWH", "SpatialWC", "SpatialAWC"),
  ...
)
}
\arguments{
\item{k}{The number of neighbors used when calculating the fuzzy
class memberships for fuzzy metrics.}

\item{metrics}{a vector of metric names to compute.}

\item{...}{Optional params for \code{\link[=getFuzzyPartitionMetrics]{getFuzzyPartitionMetrics()}}.}
}
\description{
Computes a selection of external clustering evaluation metrics for spatial
data at the class/cluster level.
}
\examples{
data <- sp_toys
getSpatialClassExternalMetrics(data$label, data$p1, data[,c("x", "y")], k=6)
getSpatialClassExternalMetrics(data$label, data$p2, data[,c("x", "y")], k=6)
}
