% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialExternalMetrics.R
\name{getSpatialClassExternalMetrics}
\alias{getSpatialClassExternalMetrics}
\title{Compute class-level external evaluation metrics for spatially-resolved data}
\usage{
getSpatialClassExternalMetrics(
  true,
  pred,
  location,
  k = 6,
  alpha = 0.5,
  metrics = c("SpatialWH", "SpatialAWH", "SpatialWC", "SpatialAWC"),
  fuzzy_true = TRUE,
  fuzzy_pred = FALSE,
  ...
)
}
\arguments{
\item{true}{A vector containing the labels of the true classes. Must be a
vector of characters, integers, numerics, or a factor, but not a list.}

\item{pred}{A vector containing the labels of the predicted clusters. Must
be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{location}{A matrix or data.frame of coordinates}

\item{k}{The number of neighbors used when calculating the fuzzy
class memberships for fuzzy metrics.}

\item{alpha}{The parameter to control to what extend the spot itself contribute
to the class composition calculation. \code{"equal"} means it is weighted the
same as other neighbors. A numeric value between \code{0} and \code{1} means the weight of the
frequency contribution for the spot itself, and the frequency contribution
for its knn is then \code{1-alpha}. By default \code{0.5}.}

\item{metrics}{a vector of metric names to compute.}

\item{...}{Optional params for \code{\link[=FuzzyPartitionMetrics]{FuzzyPartitionMetrics()}} or \code{\link[=findSpatialKNN]{findSpatialKNN()}}.}
}
\description{
Computes a selection of external clustering evaluation metrics for spatial
data at the class/cluster level.
}
\examples{
data <- sp_toys
getSpatialClassExternalMetrics(data$label, data$p1, data[,c("x", "y")], k=6)
getSpatialClassExternalMetrics(data$label, data$p2, data[,c("x", "y")], k=6)
}
