% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialExternalMetrics.R
\name{getSpatialElementExternalMetrics}
\alias{getSpatialElementExternalMetrics}
\title{getSpatialElementExternalMetrics}
\usage{
getSpatialElementExternalMetrics(
  true,
  pred,
  location,
  k = 6,
  alpha = 0.5,
  metrics = c("spotAgreement"),
  fuzzy_true = TRUE,
  fuzzy_pred = FALSE,
  ...
)
}
\arguments{
\item{k}{The number of neighbors used when calculating the fuzzy
class memberships for fuzzy metrics, or when calculating the weighted
accuracy.}

\item{metrics}{a vector of metric names to compute.}

\item{fuzzy_true}{Logical. Indicating if the \code{true} label will be transformed
into fuzzy cluster membership representation.}

\item{fuzzy_pred}{Logical. Indicating if the \code{pred} label will be transformed
into fuzzy cluster membership representation.}

\item{...}{Optional params for \code{\link[=getFuzzyPartitionElementMetrics]{getFuzzyPartitionElementMetrics()}} or \code{\link[=findSpatialKNN]{findSpatialKNN()}}.}
}
\description{
Computes a selection of external clustering evaluation metrics for spatial
data at the element level.
}
\examples{
data <- sp_toys
getSpatialElementExternalMetrics(data$label, data$p1, data[,c("x", "y")], k=6)
}
