% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialInternalMetrics.R
\name{getSpatialInternalMetrics}
\alias{getSpatialInternalMetrics}
\title{Compute internal metrics for spatial data}
\usage{
getSpatialInternalMetrics(
  labels,
  location,
  k = 6,
  level = "class",
  metrics = c("CHAOS", "PAS", "ELSA"),
  ...
)
}
\arguments{
\item{labels}{A vector containing the labels to be evaluated.}

\item{location}{A numerical matrix containing the location information, with
rows as samples and columns as location dimensions.}

\item{k}{The size of the spatial neighborhood to look at for each spot.
This is used for calculating PAS and ELSA scores.}

\item{level}{The level to calculate the metrics. Options include \code{"element"},
\code{"class"} and \code{"dataset"}.}

\item{metrics}{The metrics to compute. See details.}

\item{...}{Optional params for \code{\link[=PAS]{PAS()}}.}
}
\value{
A data.frame of metrics.
}
\description{
Computes a selection of internal clustering evaluation metrics for spatial
data.
}
\details{
The allowed values for \code{metrics} depend on the value of \code{level}:
\itemize{
\item If \code{level = "element"}, the allowed \code{metrics} are: \code{"PAS"}, \code{"ELSA"}.
\item If \code{level = "class"}, the allowed \code{metrics} are: \code{"CHAOS"}, \code{"PAS"}, \code{"ELSA"}.
\item If \code{level = "dataset"}, the allowed \code{metrics} are:
\itemize{
\item \code{"PAS"}: Proportion of abnormal spots (PAS score)
\item \code{"ELSA"}: Entropy-based Local indicator of Spatial Association (ELSA score)
\item \code{"CHAOS"}: Spatial Chaos Score.
\item \code{"MPC"}: Modified partition coefficient
\item \code{"PC"}: Partition coefficient
\item \code{"PE"}: Partition entropy
}
}
}
\examples{
data(sp_toys)
data <- sp_toys
getSpatialInternalMetrics(data$label, data[,c("x", "y")], k=6, level="class")
}
