% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noisy_moon}
\alias{noisy_moon}
\title{The noisy moon dataset
A simple toy dataset consists of two interleaving half circles.}
\format{
\subsection{\code{noisy_moon}}{

A data frame with 100 rows and 5 columns:
\describe{
\item{x, y}{Coordinates of each observations.}
\item{label}{Ground truth labels. Either 1 or 2.}
\item{kmeans_label}{Predicted clustering labels using kmeans with 2 centers.}
\item{hdbscan_label}{Predicted clustering labels using hdbscan with \code{minPts = 5}.}
}
}
}
\usage{
noisy_moon
}
\description{
The noisy moon dataset
A simple toy dataset consists of two interleaving half circles.
}
\keyword{datasets}
